+++ dbt/macros/mapa.sql
@@ -1,0 +1,3 @@
+{% macro mapa_fonte_oficial() %}
+  https://dados.agricultura.gov.br/dataset/tabua-de-risco-zoneamento-agricola-de-risco-climatico
+{% endmacro %}